Added a setup file to the vitest CONFIGURation, so that certain built-in functions that I always have to stub - like useRouter() - gets stubbed automatically
