+++ frontend/components/modules/Editor/EditorDropdownAddTags.test.ts
@@ -27,6 +27,7 @@ describe('Testing the add tags dropdown', () => {
         })
     });
 
+
     afterEach(() => {
         if (wrapper) {
             wrapper.unmount();
