+++ frontend/tests/components/modules/Blogg/FeedPagination.test.ts
@@ -1,30 +1,32 @@
-import { VueWrapper, mount } from '@vue/test-utils'
 import { createTestingPinia } from '@pinia/testing'
 import FeedPagination from '~/components/modules/Blogg/FeedPagination.vue'
+import { shallowMount } from '@vue/test-utils'
 
-let wrapper:VueWrapper
-let store
+
+
 let paginationStore:any
-let pinia
+const pinia = createTestingPinia()
+let wrapper: any;
 
+const factory = () => {
+    return shallowMount(FeedPagination, {
+        global: {
+            plugins: [],
+            components: {},
+            mocks: {},
+            stubs: {},
+        },
+        props: {},
+        slots: {}
+    })
+};
 describe("testing Pagniation component", () => {
 
     beforeEach(() => {
-        pinia = createTestingPinia()
-        store = useGeneralStore()
-        paginationStore = usePaginationStore()
+        paginationStore = usePaginationStore(pinia)
 
-        paginationStore.number_of_posts = 78
-        paginationStore.all_pages_count = 28
-        paginationStore.previous_page = "http://example.com/posts/page?=3"
-        paginationStore.current_page_number = 4
-        paginationStore.next_page = "http://example.com/posts/page?=5"
-        
-        wrapper = mount(FeedPagination, {
-            globals: {
-                plugins: [pinia]
-            }
-        })
+        paginationStore.currentPageNumber = 1
+        paginationStore.countOfPosts
 
     })
 
@@ -32,44 +34,16 @@ describe("testing Pagniation component", () => {
         wrapper.unmount()
     })
 
-    test("paginationStore variables are being rendered", () => {
-        
-        expect(wrapper.text()).toContain(78)
-        expect(wrapper.text()).toContain(28)
-        expect(wrapper.text()).toContain(4)
+    test('Should exist', () => {
+        wrapper = factory()
 
+        expect(wrapper.exists()).toBe(true)
     })
-    test("Should display navigation if there are 11 or more posts", async () => {
-
-        const navigation = wrapper.find("[data-test='post-navigation']")
-
-        expect(navigation.exists()).toBe(true)
+    test('Should match snapshot', () => {
+        wrapper = factory()
+        expect(wrapper.html()).toMatchSnapshot()
     })
-    test("Should not display navigation if there are only 10 or less posts", async () => {
-        paginationStore.number_of_posts = 10
+    
 
-        await wrapper.vm.$nextTick()
 
-        const navigation = wrapper.find("[data-test='post-navigation']")
-
-        expect(navigation.exists()).toBe(false)
-    })
-    test("Should not display navigation when number of posts is 9", async () => {
-        paginationStore.number_of_posts = 9
-
-        await wrapper.vm.$nextTick()
-
-        const navigation = wrapper.find("[data-test='post-navigation']")
-
-        expect(navigation.exists()).toBe(false)
-    })
-    test("Should display navigation if there are 11 posts", async () => {
-        paginationStore.number_of_posts = 11
-
-        await wrapper.vm.$nextTick()
-
-        const navigation = wrapper.find("[data-test='post-navigation']")
-
-        expect(navigation.exists()).toBe(true)
-    })
 })
