+++ frontend/tests/setup.ts
@@ -1,0 +1,16 @@
+import { beforeEach, vi } from "vitest"
+
+beforeEach(() => {
+    vi.stubGlobal("useRoute", () => {
+        return null
+    })
+    vi.stubGlobal("useRouter", () => {
+        return null
+    })
+    vi.stubGlobal("defineOptions", () => {
+        return null
+    })
+    vi.stubGlobal("definePageMeta", () => {
+        return null
+    })
+})
